% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{dataset}
\alias{dataset}
\title{Sensibilidad estacional}
\format{
Un data frame con 358 filas y 18 variables:
\itemize{
\item \code{id}: Número identificador.
\item \code{genero}: Género del o la participante, Masculino o Femenino.
\item \code{edad}: Factor. Edad en años, 18-35 años o > 35 años.
\item \code{deporte}: Factor. Realiza deporte, Sí o no.
\item \code{deporte_dias_semana}: Factor. Cuantas veces a la semana realiza deporte, 1 vez,
2 veces o ≥ 3 veces.
\item \code{deporte_intensidad}: Factor. Con que intensidad entrenas, Baja, Media o Alta.
\item \code{deporte_minutos_sesion}: Duración de la sesión en minutos.
\item \code{ss_patron_verano}: Patrón de verano, unidad adimensional.
\item \code{ss_patron_invierno}: Patrón de invierno, unidad adimensional.
\item \code{ss_patron_tipo}: Factor. Tipo de patrón estacional, Verano, Invierno o Mixto.
\item \code{ss_index}: Factor. Índice de puntaje estacional (SSI), Normal, Winter blues o SAD.
\item \code{ss_severidad}: Factor. Severidad del patrón estacional, No es problema, Leve,
Moderado, Importante, Severo o Grave.
\item \code{riff_*}: Dominios del cuestionario de bienestar de Ryff. Entre los dominios se
encuentran la 'Autoaceptacion' (\code{riff_autoaceptacion}), 'Relaciones positivas' (\code{riff_relaciones_positivas}),
'Autonomía' (\code{riff_autonomia}), 'Dominio del entorno' (\code{riff_dominio_entorno}), 'Crecimiento personal'
(\code{riff_crecimiento_personal}), así como 'Propósito en la vida' (\code{riff_proposito_vida}).
}
}
\source{
Los datos son parte del estudio de sensibilidad estacional llevado a cabo por Matías Castillo-Aguilar y Cristián Núñez-Espinosa.
}
\usage{
dataset
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Datos del estudio de sensibilidad estacional que contiene variables sociodemográficas,
sensibilidad estacional, y de bienestar evaluadas mediante el cuestionario de bienestar
de Ryff
}
\keyword{datasets}
