% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_normal.R
\name{is_normal}
\alias{is_normal}
\title{Comprueba si un vector numérico sigue una distribución normal}
\usage{
is_normal(x, alpha = 0.05, test = NULL)
}
\arguments{
\item{x}{Un vector numérico.}

\item{alpha}{El valor de alfa para el rechazo de la hipótesis nula de normalidad. Por defecto es 0.05.}

\item{test}{Una función que acepte a \code{x} y devuelva un valor-p para ser contrastado con \code{alpha}. Por
defecto se escoge basado en el tamaño muestral.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Basado en el tamaño muestral, comprueba analíticamente si un vector numérico
sigue una distribución normal, para el rechazo de la hipótesis nula usa un alfa
del 0.05 o 5\%. Si el tamaño muestral es menor o igual a 50 usa la prueba de
Shapiro-Wilk, y en caso contrario la prueba de Kolmogorov-Smirnov con corrección
de Lilliefors del paquete \code{nortest}.
}
