% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_test.R
\name{ks_test}
\alias{ks_test}
\title{Prueba de normalidad de Lilliefors (Kolmogorov-Smirnov)}
\usage{
ks_test(x)
}
\arguments{
\item{x}{A numeric vector to be tested.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Realiza la prueba de Lilliefors (Kolmogorov-Smirnov) para la hipótesis de normalidad.
Para más detalles ver \code{nortest::lillie.test()}.
}
