% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_colwise.R
\name{n_colwise}
\alias{n_colwise}
\title{Cuantos valores (sin NA) hay por columna}
\usage{
n_colwise(.data, .cols = "all", .by, char.only = FALSE)
}
\arguments{
\item{.data}{Un data.table. Si no lo es debe ser al menos un data.frame y despues es convertido a un data.table.}

\item{.cols}{Vector de nombre de columnas o función que devuelva un valor lógico evaluando la columna (e.g., is.numeric).}

\item{.by}{Una variable en base a la cual agrupar los análisis, por defecto no se usa ninguno.}

\item{char.only}{Lógico. El argumento \code{.by} debe tratarse como un carácter? Por defecto es \code{FALSE}. De lo contrario se trata como un nombre sin comillas.}
}
\value{
data.table con tamaños muestrales por cada columna y grupos especificados.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Obten el tamaño muestral por cada columna de un data.table (es convertido a uno si no lo es), sin
contar aquellas observaciones con valores perdidos
}
