% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_not_na.R
\name{n_not_na}
\alias{n_not_na}
\title{Cuenta el número de elementos que no son valores perdidos}
\usage{
n_not_na(x)
}
\arguments{
\item{x}{Un vector numérico.}
}
\value{
Número entero.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Esta función suma la cantidad de valores que no son NA, utilizando
un algoritmo sencillo y fácil de implementar.
}
