% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalSensitivity-package.R
\docType{package}
\name{seasonalSensitivity-package}
\alias{seasonalSensitivity}
\alias{seasonalSensitivity-package}
\title{seasonalSensitivity: Data and analysis from the seasonal sensitivity study}
\description{
Data and analysis from the seasonal sensitivity study.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NIM-ACh/seasonalSensitivity}
  \item \url{https://nim-ach.github.io/seasonalSensitivity}
  \item Report bugs at \url{https://github.com/NIM-ACh/seasonalSensitivity/issues}
}

}
\author{
\strong{Maintainer}: Matías Castillo Aguilar \email{matcasti@umag.cl} (\href{https://orcid.org/0000-0001-7291-247X}{ORCID})

Authors:
\itemize{
  \item Cristian Núñez Espinosa \email{cristian.nunez@umag.cl} (\href{https://orcid.org/0000-0002-9896-7062}{ORCID}) [data contributor]
}

}
\keyword{internal}
